requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.30
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
faiss-cpu>=1.7.4
pypdf>=3.17.0
python-dotenv>=1.0.0
tiktoken>=0.5.0

# Optional (for future extensions)
fastapi>=0.110.0
uvicorn>=0.29.0